Move Mix.* modules to Mix.Tasks.Hex
