+++ test/support/case.ex
@@ -153,7 +153,7 @@ defmodule HexTest.Case do
 
   def setup_auth(username, password) do
     {201, body, _} = Hex.API.Key.new("setup_auth", [user: username, pass: password])
-    Mix.Hex.Utils.persist_key(password, body["secret"])
+    Mix.Tasks.Hex.persist_key(password, body["secret"])
   end
 
   def get_auth(username, password) do
